pyproject.toml
[project]
name = "ras545"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.69.0",
    "celery>=5.5.3",
    "langchain>=1.0.2",
    "notebook>=7.4.5",
    "open-clip-torch>=3.2.0",
    "openai>=2.6.0",
    "pillow>=11.3.0",
    "pydantic>=2.11.9",
    "pydobot>=1.3.2",
    "pydobot2>=0.1.0",
    "python-dotenv>=1.1.1",
    "redis>=7.0.1",
    "scikit-image>=0.25.2",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "ultralytics>=8.3.206",
    "websocket-client>=1.8.0",
]
